requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.3.3
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-taggit==1.3.0
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
html5lib==1.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.6
pylint==2.4.4
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-decouple==3.4
pytz==2020.1
requests==2.24.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.11
virtualenv==20.0.30
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.11.2
